requirements.txt
Django>=4.2,<5.0
Pillow>=10.2.0
django-ckeditor>=6.7.1
Markdown
django-rosetta>=0.9.9
django-admin-honeypot-updated-2021>=1.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
python-decouple>=3.8
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
